Move pipes tests into Modulos.Test project.
