+++ tests/Modulos.Tests/Modulos.Tests.csproj
@@ -12,11 +12,14 @@
 
   <ItemGroup>
     <Compile Include="Tests\.Unit\AssemblyExploringTests.cs" />
+    <Compile Include="Tests\.Unit\Pipes\Pipeline_execution_tests.cs" />
+    <Compile Include="Tests\.Unit\Pipes\Pipeline_modify_collection_tests.cs" />
   </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="FluentAssertions" Version="5.10.3" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.0" />
+    <PackageReference Include="Moq" Version="4.15.1" />
     <PackageReference Include="xunit" Version="2.4.1" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
       <PrivateAssets>all</PrivateAssets>
