+++ tests/Modulos.Tests/Tests/.Unit/Pipes/Pipeline_execution_tests.cs
@@ -9,7 +9,7 @@ using Xunit;
 // ReSharper disable InconsistentNaming
 // ReSharper disable ClassNeverInstantiated.Local
 
-namespace Modulos.Messaging.Tests.Tests.Unit
+namespace Modulos.Tests.Unit.Pipes
 {
     public class Pipeline_execution_tests
     {
